+++ CHANGELOG.markdown
@@ -5,6 +5,7 @@
 - Add Vibe/ConsecutiveConstantAlignment cop.
 - Add Vibe/ConsecutiveLetAlignment cop.
 - Add Vibe/RspecStubChainStyle cop.
+- Add AllowedCops configuration to Vibe/NoRubocopDisable cop.
 
 ## 0.1.0 — January 6th, 2026